KU2
The AImove script is attached to the AI block which is to move to the target. The seeker is referenced to generate the path and and an instance of it generated as well for the path and a path is generated using the position of the AI block annd the targets position. Then in the start method the MoveTowardsTarget coroutine is called with the position of this.gameobject as its parameters. Firstly a vector3 list is created with the vector path and using a counter is looped throughout each position. A while loop is looped if the distance between the block and next position is larger or equal to 0.5. The block moves one unit towards the position in the list and then the path is refresed with new coordinates just in case that the target moves in future versions. 0.1 seconds are passed until this is looped again until all positions in the list have been visited.