KU2-b
Firstly a public gameobject of obstacle is referenced. Then a generateObstacles method is created.  Random X and Y coordinates within the range of the camera view are generated and created into a vector3. This position is then used to instaniate the gameobject in the random position. This is looped 5 times in the start method as required. Then next in the start method the referenced graphParent is used to scan the  grid with the new obstacles. The grid is scanned only once after the obstacles are inistaniated, if needed to update continously, a coroutine is needed.